cpython: tolerate bindgen ffi differences
User request: evaluate whether rust-cpython can compile against the new experimental bindgen mode in python3-sys.

I updated cpython's call sites and macros to accept the bindgen-produced signatures and integer widths while preserving the existing handwritten-binding behavior. The core decision was to keep the adaptation narrow and feature-gated: cast constants where bindgen emits narrower integer types, unwrap optional callback pointers only where bindgen models nullable C function pointers, and teach method-definition macros to handle bindgen's PyCFunction representation so the library and test targets compile without broad API churn.

Agent-Prompt: 我在 python3-sys 新增了实验性的 bindgen 功能。帮我看看 rust-cpython 和 bindgen python3-sys 兼容性如何，能不能编译
Agent-Model: gpt-5.4

+++ src/buffer.rs
@@ -167,7 +167,11 @@ impl PyBuffer {
             let mut buf = Box::new(mem::zeroed::<ffi::Py_buffer>());
             err::error_on_minusone(
                 py,
-                ffi::PyObject_GetBuffer(obj.as_ptr(), &mut *buf, ffi::PyBUF_FULL_RO),
+                ffi::PyObject_GetBuffer(
+                    obj.as_ptr(),
+                    &mut *buf,
+                    ffi::PyBUF_FULL_RO as libc::c_int,
+                ),
             )?;
             validate(&buf);
             Ok(PyBuffer(buf))
